requirements.txt
Django==4.2.7
mysqlclient==2.2.0
python-decouple==3.8
django-extensions==3.2.3
django-crispy-forms==2.1
crispy-bootstrap5==0.7
Pillow==10.1.0
openpyxl==3.1.2
reportlab==4.0.7
python-dateutil==2.8.2
psycopg2-binary==2.9.9